# contracts/openapi/learning.openapi.yaml
openapi: 3.0.3
info:
  title: Learning Service (Impilo Fundo) — Internal API
  description: >
    Internal API surface for **Impilo Fundo**, the **native learning
    management, certification, in-service training, pre-service training
    and CPD support capability of Impilo vNext**. Fundo is designed to
    operate as a complete modern LMS at national scale — its own
    catalogue, paths, enrolment, progress, assessments, certificates,
    learning records and analytics — without depending on any external
    LMS to function.


    External LMS platforms (Moodle, Open edX, Canvas, Chamilo, …) may be
    connected through **optional adapters** where useful, but Fundo must
    not depend on any external LMS. The existing Moodle orchestration
    endpoints documented below are one such optional adapter, preserved
    verbatim for backwards compatibility — not the Fundo runtime.


    This contract is **descriptive of the current implementation only** —
    it documents the endpoints that exist in `services/learning-service`
    today and does not assert new behaviour, new fields, or future
    surfaces. Endpoints under `/internal/v1/learning/integrations/**`
    and `/internal/v1/learning/orchestration/**` are server-to-server
    only and use shared internal keys (see `securitySchemes`), not
    end-user JWT.


    CPD authority remains with `varapi-service`. The Impilo Fundo brand
    is the product/UX name; `learning-service` is its technical home.
  version: 0.1.0
...
servers:
...
paths:
  /internal/v1/learning/workflow-context:
    get:
      tags: [Internal]
      summary: Resolve the workflow surface for a given app / route / roles
      description: >
        Returns the contextual learning surface (resources, hints, role
        requirements) for the supplied app/route/workflow tuple. Source for
        the contextual learning panels in the One Experience Shell.
      operationId: getWorkflowContext
      parameters:
        - { name: appCode, in: query, required: true, schema: { type: string } }
        - { name: routeRef, in: query, required: true, schema: { type: string } }
        - { name: workflowCode, in: query, required: false, schema: { type: string } }
        - { name: roles, in: query, required: false, description: "Comma-separated role codes", schema: { type: string } }
        - { name: subjectType, in: query, required: false, schema: { type: string } }
        - { name: subjectId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Workflow surface resolved
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_Map' }
  /internal/v1/learning/helpdesk/{issueType}:
    get:
      tags: [Internal]
      summary: List helpdesk knowledge-base entries for an issue type or topic
      operationId: listHelpdeskKnowledge
      parameters:
        - { name: issueType, in: path, required: true, schema: { type: string } }
        - { name: subjectType, in: query, required: false, schema: { type: string } }
        - { name: subjectId, in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Helpdesk items
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_ItemsList' }
...
  /internal/v1/learning/resource-opened:
    post:
      tags: [Internal]
      summary: Record that a user opened a learning resource (audit + progress)
      operationId: recordResourceOpened
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResourceOpenedRequest' }
      responses:
        '200':
          description: Event recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_Status' }
  /internal/v1/learning/resources/{resourceId}:
    get:
      tags: [Internal]
      summary: Get the public view of a learning resource by id
      operationId: getResourceById
      parameters:
        - { name: resourceId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Resource public view
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_Map' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /internal/v1/learning/subject-completions:
    get:
      tags: [Internal]
      summary: List learning completions for a subject (user / provider / etc.)
      operationId: listSubjectCompletions
      parameters:
        - { name: subjectType, in: query, required: true, schema: { type: string } }
        - { name: subjectId, in: query, required: true, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 25 } }
      responses:
        '200':
          description: Completion rows
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_ItemsList' }
  /internal/v1/learning/v11/subject-completions:
    get:
      tags: [InternalV11]
      summary: List learning completions for a subject — native v1.1 read surface (Phase 3C)
      description: >
        Additive **v1.1-compliant** read endpoint for Impilo Fundo. Delegates to
        the same native facade as the legacy
        `GET /internal/v1/learning/subject-completions` and never invokes
        Moodle or any external LMS. Enforces the platform v1.1 contract
        (companion header enforcement, error envelope, client-timeout
        enforcement) via the shared tech-companion filter chain. The legacy
        endpoint remains unchanged.


        When `subjectType` / `subjectId` are omitted, or when the
        `X-Tenant-ID` header is not a valid UUID, the endpoint returns a
        successful empty page (HTTP 200 with `data.items = []`) rather than
        a 4xx — the surface is intentionally read-only and side-effect-free.
      operationId: listSubjectCompletionsV11
      parameters:
        - { name: subjectType, in: query, required: false, schema: { type: string } }
        - { name: subjectId, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 25 } }
      responses:
        '200':
          description: Completion rows (v1.1 envelope; may be empty)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_ItemsList' }
        '400':
          description: Required v1.1 header missing (`X-Tenant-ID`, `X-Pod-ID`, `X-Request-ID`, `X-Correlation-ID`)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '504':
          description: Client timeout exceeded (`X-Client-Timeout-MS` reached)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /internal/v1/learning/subject-profile:
    post:
      tags: [Internal]
      summary: Upsert a subject's learning profile (link/update Fundo user reference and metadata)
      operationId: upsertSubjectProfile
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubjectProfileUpsertRequest' }
      responses:
        '200':
          description: Profile upserted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiEnvelope_Map' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Used by the default security chain when `learning.security.oauth2-enabled=true`.
        Compose dev environments run with this disabled (see
        `ops/runtime/docker-compose.operations.yml`); Helm defaults to true.
    integrationKey:
      type: apiKey
      in: header
      name: X-Impilo-Learning-Internal-Key
      description: >
        Shared internal key for trusted ingress from `varapi-service` after a
        Fundo webhook creates a governed CPD candidate.
    orchestrationKey:
      type: apiKey
      in: header
      name: X-Impilo-Learning-Orchestration-Key
      description: >
        Shared internal key for server-to-server orchestration jobs (Tshepo /
        Varapi schedulers, Moodle Web Service relays).
  schemas:
...
    ApiEnvelope_Map:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
      required: [data]
    ApiEnvelope_ItemsList:
      type: object
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                type: object
                additionalProperties: true
          required: [items]
      required: [data]
...
    ApiEnvelope_Status:
      type: object
      properties:
        data:
          type: object
          properties:
            status: { type: string }
          required: [status]
      required: [data]
...
    Error:
      type: object
      properties:
        error: { type: string }

    # ── Request bodies ────────────────────────────────────────────────────
    ResourceOpenedRequest:
      type: object
      properties:
        resourceId: { type: string, format: uuid }
        workflowContext:
          type: object
          additionalProperties: true
      required: [resourceId]
    SubjectProfileUpsertRequest:
      type: object
      description: >
        `fundoUserRef` and `metadata` are both **patch-style**: presence of
        the key triggers an update (including to `null`); absence leaves the
        existing value unchanged.
      properties:
        subjectType: { type: string }
        subjectId: { type: string }
        fundoUserRef: { type: string, nullable: true }
        metadata:
          type: object
          additionalProperties: true
      required: [subjectType, subjectId]
...